Create README.md
This commit creates a README file for the notes app. The README file provides information about the app, its usage, and its license.

The README file includes the following sections:

Overview: This section provides a brief overview of the notes app, including its purpose and features.
Usage: This section provides instructions on how to use the notes app, including how to add, remove, list, and read notes.
License: This section provides the license under which the notes app is distributed.
The README file is written in Markdown, which is a lightweight markup language that is easy to read and write.

This commit is important because it makes the notes app more user-friendly. The README file provides users with all the information they need to start using the app immediately.

Additional information
In addition to the above, you may also want to include the following information in your extended description:

The motivation for creating the README file. For example, you may want to mention that you created the README file to make the app more user-friendly, or to provide users with more information about the app's features.
Any changes that you made to the README file. For example, you may want to mention that you added a new section to the README file, or that you updated an existing section.
Any known issues with the README file. For example, you may want to mention that a particular section of the README file is incomplete, or that a particular command is not working as expected.